+++ src/lega/frontend.py
@@ -34,9 +34,23 @@ import aiohttp_jinja2
 
 from .conf import CONF
 from .utils import db
-from .utils import only_central_ega
 
 LOG = logging.getLogger('frontend')
+
+def only_central_ega(async_func):
+    '''Decorator restrain endpoint access to only Central EGA'''
+    @wraps(async_func)
+    async def wrapper(request):
+        # Just an example
+        if request.headers.get('X-CentralEGA', 'no') != 'yes':
+            raise web.HTTPUnauthorized(text='Not authorized. You should be Central EGA.\n')
+        # Otherwise, it is from CentralEGA, we continue
+        res = async_func(request)
+        res.__name__ = getattr(async_func, '__name__', None)
+        res.__qualname__ = getattr(async_func, '__qualname__', None)
+        return (await res)
+    return wrapper
+
 
 @aiohttp_jinja2.template('index.html')
 async def index(request):
